Split `boxed.rs` into a few modules
+ some minor style changes
